+++ src/main/java/com/example/springrider/domain/cart/dto/request/CreateCartItemBulkRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.springrider.domain.cart.dto;
+package com.example.springrider.domain.cart.dto.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
