+++ src/main/java/com/example/springrider/domain/cart/dto/request/CreateCartItemRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.springrider.domain.cart.dto;
+package com.example.springrider.domain.cart.dto.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
