+++ src/main/java/com/example/springrider/domain/cart/dto/response/CreateCartItemBulkResponseDto.java
@@ -1,5 +1,7 @@
-package com.example.springrider.domain.cart.dto;
+package com.example.springrider.domain.cart.dto.response;
 
+import com.example.springrider.domain.cart.dto.CreateFailedDto;
+import com.example.springrider.domain.cart.dto.CreateSuccessDto;
 import java.util.List;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
