+++ src/main/java/com/example/springrider/domain/cart/service/CartService.java
@@ -1,23 +1,23 @@
 package com.example.springrider.domain.cart.service;
 
-import com.example.springrider.domain.cart.dto.CreateCartItemBulkRequestDto;
-import com.example.springrider.domain.cart.dto.CreateCartItemBulkResponseDto;
-import com.example.springrider.domain.cart.dto.CreateCartItemRequestDto;
 import com.example.springrider.domain.cart.dto.CreateFailedDto;
 import com.example.springrider.domain.cart.dto.CreateSuccessDto;
-import com.example.springrider.domain.cart.dto.FindCartItemBulkResponseDto;
-import com.example.springrider.domain.cart.dto.FindCartItemResponseDto;
-import com.example.springrider.domain.cart.dto.UpdateCartItemRequestDto;
-import com.example.springrider.domain.cart.dto.UpdateCartItemResponseDto;
+import com.example.springrider.domain.cart.dto.request.CreateCartItemBulkRequestDto;
+import com.example.springrider.domain.cart.dto.request.CreateCartItemRequestDto;
+import com.example.springrider.domain.cart.dto.request.UpdateCartItemRequestDto;
+import com.example.springrider.domain.cart.dto.response.CreateCartItemBulkResponseDto;
+import com.example.springrider.domain.cart.dto.response.FindCartItemBulkResponseDto;
+import com.example.springrider.domain.cart.dto.response.FindCartItemResponseDto;
+import com.example.springrider.domain.cart.dto.response.UpdateCartItemResponseDto;
 import com.example.springrider.domain.cart.entity.CartItem;
 import com.example.springrider.domain.cart.enums.CartItemStatus;
 import com.example.springrider.domain.cart.repository.CartRepository;
-import com.example.springrider.global.exception.ExceptionCode;
-import com.example.springrider.global.exception.InvalidRequestException;
 import com.example.springrider.domain.menu.entity.Menu;
 import com.example.springrider.domain.menu.repository.MenuRepository;
 import com.example.springrider.domain.user.entity.User;
 import com.example.springrider.domain.user.repository.UserRepository;
+import com.example.springrider.global.exception.ExceptionCode;
+import com.example.springrider.global.exception.InvalidRequestException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
